# docs/api/clubspot.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ClubSpot.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/agenda": {
      "get": {
        "tags": [
          "bookings"
        ],
        "operationId": "GetAgenda",
        "parameters": [
          {
            "name": "sport",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agenda"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/bookings": {
      "post": {
        "tags": [
          "bookings"
        ],
        "operationId": "CreateBooking",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingCreatedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/api/bookings/{id}/cancel": {
      "post": {
        "tags": [
          "bookings"
        ],
        "operationId": "CancelBooking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/bookings/{id}/checkout": {
      "post": {
        "tags": [
          "bookings"
        ],
        "operationId": "CreateBookingCheckout",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingCheckoutResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/people/{id}/bookings": {
      "get": {
        "tags": [
          "bookings"
        ],
        "operationId": "GetPersonBookings",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonBookingResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/portal/{clubSlug}/bookings/{id}": {
      "get": {
        "tags": [
          "portal"
        ],
        "operationId": "GetPortalBooking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "clubSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingSnapshot"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/portal/{clubSlug}/bookings": {
      "post": {
        "tags": [
          "portal"
        ],
        "operationId": "CreatePortalBooking",
        "parameters": [
          {
            "name": "clubSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalBookingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalBookingCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
...
    "/api/payments/fake/webhook/{clubSlug}": {
      "post": {
        "tags": [
          "payments"
        ],
        "operationId": "PostFakeWebhook",
        "parameters": [
          {
            "name": "clubSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FakeWebhookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentApplyResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/payments/mercadopago/webhook/{clubSlug}": {
      "post": {
        "tags": [
          "payments"
        ],
        "operationId": "PostMercadoPagoWebhook",
        "parameters": [
          {
            "name": "clubSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Agenda": {
        "required": [
          "currency",
          "courts",
          "inactive"
        ],
        "type": "object",
        "properties": {
          "currency": {
            "type": "string"
          },
          "courts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgendaCourt"
            }
          },
          "inactive": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgendaInactiveBooking"
            }
          }
        }
      },
      "AgendaBooking": {
        "required": [
          "id",
          "startMinute",
          "durationMinutes",
          "customerName",
          "customerPhone",
          "price",
          "paidAmount",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "startMinute": {
            "type": "integer",
            "format": "int32"
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "customerName": {
            "type": "string"
          },
          "customerPhone": {
            "type": [
              "null",
              "string"
            ]
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "paidAmount": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          }
        }
      },
      "AgendaCourt": {
        "required": [
          "courtId",
          "name",
          "detail",
          "isCovered",
          "windows",
          "slots",
          "bookings"
        ],
        "type": "object",
        "properties": {
          "courtId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "isCovered": {
            "type": "boolean"
          },
          "windows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimeRange"
            }
          },
          "slots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgendaSlot"
            }
          },
          "bookings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgendaBooking"
            }
          }
        }
      },
      "AgendaInactiveBooking": {
        "required": [
          "id",
          "courtId",
          "courtName",
          "startMinute",
          "durationMinutes",
          "customerName",
          "customerPhone",
          "price",
          "paidAmount",
          "status",
          "cancelledAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "courtId": {
            "type": "string",
            "format": "uuid"
          },
          "courtName": {
            "type": "string"
          },
          "startMinute": {
            "type": "integer",
            "format": "int32"
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "customerName": {
            "type": "string"
          },
          "customerPhone": {
            "type": [
              "null",
              "string"
            ]
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "paidAmount": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "cancelledAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "AgendaSlot": {
        "required": [
          "startMinute",
          "duration",
          "price"
        ],
        "type": "object",
        "properties": {
          "startMinute": {
            "type": "integer",
            "format": "int32"
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "BookingCheckoutResponse": {
        "required": [
          "url",
          "amount",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BookingCreatedResponse": {
        "required": [
          "id",
          "price"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "BookingRequest": {
        "required": [
          "courtId",
          "date",
          "startMinute",
          "durationMinutes",
          "customerName",
          "customerPhone"
        ],
        "type": "object",
        "properties": {
          "courtId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "startMinute": {
            "type": "integer",
            "format": "int32"
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "customerName": {
            "type": "string"
          },
          "customerPhone": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "BookingSnapshot": {
        "required": [
          "id",
          "courtId",
          "courtName",
          "sport",
          "date",
          "startMinute",
          "durationMinutes",
          "price",
          "paidAmount",
          "status",
          "paymentMode",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "courtId": {
            "type": "string",
            "format": "uuid"
          },
          "courtName": {
            "type": "string"
          },
          "sport": {
            "$ref": "#/components/schemas/Sport"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "startMinute": {
            "type": "integer",
            "format": "int32"
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "paidAmount": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "paymentMode": {
            "$ref": "#/components/schemas/PaymentMode"
          },
          "expiresAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "BookingStatus": {
        "enum": [
          "confirmed",
          "cancelled",
          "pendingPayment",
          "expired"
        ],
        "type": "string"
      },
...
      "FakeWebhookRequest": {
        "required": [
          "bookingId",
          "externalId",
          "approved",
          "amount"
        ],
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string",
            "format": "uuid"
          },
          "externalId": {
            "type": "string"
          },
          "approved": {
            "type": "boolean"
          },
          "amount": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PaymentApplyOutcome": {
        "enum": [
          "confirmed",
          "rejected",
          "alreadyProcessed",
          "orphaned",
          "unknownBooking"
        ],
        "type": [
          "null",
          "string"
        ]
      },
      "PaymentApplyResponse": {
        "required": [
          "outcome"
        ],
        "type": "object",
        "properties": {
          "outcome": {
            "$ref": "#/components/schemas/PaymentApplyOutcome"
          }
        }
      },
      "PaymentMode": {
        "enum": [
          "club",
          "onlineFull",
          "onlineDeposit"
        ],
        "type": "string"
      },
...
      "PersonBookingResponse": {
        "required": [
          "id",
          "date",
          "startMinute",
          "durationMinutes",
          "courtName",
          "sport",
          "price",
          "paid",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "startMinute": {
            "type": "integer",
            "format": "int32"
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "courtName": {
            "type": "string"
          },
          "sport": {
            "$ref": "#/components/schemas/Sport"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "paid": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          }
        }
      },
...
      "PortalBookingCreatedResponse": {
        "required": [
          "id",
          "price",
          "chargeAmount",
          "status",
          "expiresAt",
          "checkoutUrl",
          "token"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "chargeAmount": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "expiresAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "checkoutUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "token": {
            "type": "string"
          }
        }
      },
      "PortalBookingRequest": {
        "required": [
          "courtId",
          "date",
          "startMinute",
          "durationMinutes",
          "customerName",
          "customerPhone",
          "customerEmail",
          "paymentMode",
          "returnUrl"
        ],
        "type": "object",
        "properties": {
          "courtId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "startMinute": {
            "type": "integer",
            "format": "int32"
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "customerName": {
            "type": "string"
          },
          "customerPhone": {
            "type": "string"
          },
          "customerEmail": {
            "type": [
              "null",
              "string"
            ]
          },
          "paymentMode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PaymentMode"
              }
            ]
          },
          "returnUrl": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "Sport": {
        "enum": [
          "padel",
          "football"
        ],
        "type": "string"
      },
      "TimeRange": {
        "required": [
          "opensAtMinute",
          "closesAtMinute"
        ],
        "type": "object",
        "properties": {
          "opensAtMinute": {
            "type": "integer",
            "format": "int32"
          },
          "closesAtMinute": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    }
  },
...
}